Updated with dylan's suggestions
