+++ src/client/features/pathways/index.js
@@ -4,7 +4,6 @@ const _ = require('lodash');
 const queryString = require('query-string');
 const Loader = require('react-loader');
 const classNames = require('classnames');
-const diceCoefficient = require('dice-coefficient');
 
 const CytoscapeService = require('../../common/cy/');
 const { ServerAPI } = require('../../services/');
@@ -14,7 +13,7 @@ const Sidebar = require('../../common/components/sidebar');
 const EmptyNetwork = require('../../common/components/empty-network');
 const InfoMenu = require('./menus/network-info-menu');
 const FileDownloadMenu = require('./menus/file-download-menu');
-const databaseURLMap = require('../../common/datasource-url-map');
+const datasourceToUrl = require('../../common/datasource-url-map');
 
 const Pathway = require('../../models/pathway/pathway-model');
 
@@ -86,31 +85,10 @@ class Pathways extends React.Component {
     }
   }
 
-  baseNameToUrl(baseName){
-    const lowercaseName = baseName.toLowerCase();
-
-    let databaseURL = databaseURLMap.get(lowercaseName);
-    if(!databaseURL){
-      const datasourceURLScores = [];
-      databaseURLMap.forEach( (datasourceURL, datasource, map) => {
-        datasourceURLScores.push([datasourceURL,diceCoefficient(datasource,lowercaseName)]);
-      });
-      datasourceURLScores.sort( (a,b) => { return b[1] - a[1]; } );
-      databaseURL = datasourceURLScores[0][0];
-    }
-
-    return databaseURL;
-  }
-
   render() {
     let { loading, pathway, cySrv, activeMenu, networkEmpty } = this.state;
 
-    const baseDatasource = pathway.datasource();
-    const databaseURL = this.baseNameToUrl(baseDatasource);
-    let baseName = pathway.name();
-    if(baseName.toLowerCase() === 'unknown title'){
-      baseName = '';
-    }
+    const databaseURL = datasourceToUrl(pathway.datasource());
 
     let network = h('div.network', { className: classNames({
       'network-loading': loading,
@@ -129,8 +107,8 @@ class Pathways extends React.Component {
       h('div.app-bar-branding', [
         h('i.app-bar-logo', { href: 'http://www.pathwaycommons.org/' }),
         h('div.app-bar-title', [
-          h('a.plain-link', { href:databaseURL,target:'_blank' }, ' ' + baseDatasource),
-          h('span', ': ' + baseName)
+          h('a.plain-link', { href:databaseURL,target:'_blank' }, ' ' + pathway.datasource()),
+          h('span', ': ' + pathway.name())
         ])
       ]),
       toolbar
